requirements.txt
# LangChain ecosystem
langchain>=0.2.16
langchain-openai>=0.1.22
langchain-community>=0.2.16
langchain-text-splitters>=0.2.4
langgraph>=0.2.28

# NVIDIA NIM free-tier fallback (used when OPENAI_API_KEY is not set)
# Get a free key at: https://build.nvidia.com/
langchain-nvidia-ai-endpoints>=0.3.5

# Vector store
faiss-cpu>=1.8.0

# LLM & embeddings
openai>=1.40.0

# Web search
tavily-python>=0.3.9

# Backend
fastapi>=0.112.0
uvicorn[standard]>=0.30.6
python-multipart>=0.0.9

# Data validation
pydantic>=2.8.0
pydantic-settings>=2.4.0

# Environment
python-dotenv>=1.0.1

# Document parsing
pypdf>=4.3.1
unstructured>=0.15.0

# HTTP client (for Streamlit → FastAPI)
requests>=2.32.3

# Frontend
streamlit>=1.38.0

# Testing
pytest>=8.3.2
httpx>=0.27.2
pytest-asyncio>=0.23.8
